+++ CHANGELOG.md
@@ -26,7 +26,8 @@ A more detailed list of changes is available in the corresponding milestones for
   - **[com.google.fonts/check/013]:** com.google.fonts/check/equal_unicode_encodings
   - **[com.google.fonts/check/014]:** com.google.fonts/check/equal_font_versions
   - **[com.google.fonts/check/015]:** com.google.fonts/check/post_table_version
-  - **[com.google.fonts/check/016]:** com.google.fonts/check/os2/fstype
+  - **[com.google.fonts/check/016]:** com.google.fonts/check/fstype
+  - **[com.google.fonts/check/018]:** com.google.fonts/check/vendor_id
   - **[com.google.fonts/check/031]:** com.google.fonts/check/name/no_copyright_on_description
   - **[com.google.fonts/check/033]:** com.google.fonts/check/monospace
   - **[com.google.fonts/check/034]:** com.google.fonts/check/xavgcharwidth